dynamic json shapes

+++ DxDispatch/src/model/Model.h
@@ -30,7 +30,7 @@ public:
         std::optional<std::string> counterName;
         uint64_t counterOffsetBytes;
 
-        std::vector<uint32_t> shape;
+        std::vector<int64_t> shape;
     };
 
     using Bindings = std::unordered_map<std::string, std::vector<BufferBindingSource>>;
